Refactor MFA management in IUserService and Auth0Client
Replaced the `ToggleMFA` method with `DisableMfa` in the `IUserService` interface and `Auth0Client` class, removing the `enable` parameter to simplify MFA management. Updated the `user.proto` file to rename the RPC method and remove the `UserMFARequest` message. Adjusted the `UserService` implementation to align with the new method signature. Ensured consistency in the API documentation by reflecting these changes in the `PublicAPI.Unshipped.txt` file.