Update query to use user_metadata.identity_id field
The `Query` property in the `GetUsersRequest` object within the
`UpdateUser` method of the `Auth0Client` class was updated to
search for the `identityId` in the `user_metadata.identity_id`
field instead of the `identity_id` field. This change aligns
with the updated data schema or corrects the query to target
the appropriate field.